Quite heafty rework, and simplification of front matter. Things are driven by structure rather than by duplicated content in front-matter.
BookToC is enabled most everywhere, it doesn't show if it's not present,
while we keep the Edit page and tweak docs links.

Side menu is not shown in `full-width` view anymore, so no need for
`sideNav: false`, and `sideNav: true` is default most everywhere.

`sideNavLinkRename` now only needed as an override to `title:`. It was
renamed to `sideNavTitle` when still in use (just a few times).

`pageTitle` renamed to `title:` which is the default for pages.
`pageTitleIcon` is now infered by the section name (eosio/ethereum/platform).

Altered the side navigation to only show the root starting at a given point,
this is now driven by `sideNavRoot: true` in the `_index.md` of the section
we want to become root.  We can choose whatever depth we want a new root.

Tweaked the side nav also to have expandable section names, instead of going
to the section page.  Therefore, removed most of the content in those `_index.md`
files, since it's not reachable anyway (except is you tweak the URL).

Moved some Authentication specifics to EOSIO in its section.

Removed some duplicate titles as a first header in the article content (duplicates
the `title:` that is already present on the page).

Wee, quite a ride.. validated each removed line across the board!